+++ verik-importer/src/main/kotlin/io/verik/importer/cast/common/CastContext.kt
@@ -20,6 +20,7 @@ import io.verik.importer.ast.sv.element.common.SvContainerElement
 import io.verik.importer.ast.sv.element.common.SvElement
 import io.verik.importer.ast.sv.element.declaration.SvDeclaration
 import io.verik.importer.ast.sv.element.declaration.SvPort
+import io.verik.importer.ast.sv.element.declaration.SvProperty
 import io.verik.importer.ast.sv.element.declaration.SvValueParameter
 import io.verik.importer.ast.sv.element.descriptor.SvDescriptor
 import io.verik.importer.ast.sv.element.expression.SvExpression
@@ -63,6 +64,10 @@ class CastContext(
             is SvContainerElement -> element.elements.map { it.cast() }
             else -> listOf(element.cast())
         }
+    }
+
+    fun castProperties(ctx: RuleContext): List<SvProperty> {
+        return castDeclarations(ctx).map { it.cast() }
     }
 
     fun castValueParameter(ctx: RuleContext): SvValueParameter? {