+++ verik-importer/src/test/kotlin/io/verik/importer/cast/cast/PropertyCasterTest.kt
@@ -49,4 +49,17 @@ internal class PropertyCasterTest : BaseTest() {
             """.trimIndent()
         ) { it }
     }
+
+    @Test
+    fun `cast property from structUnionMember`() {
+        driveCasterTest(
+            SystemVerilogParser.DataDeclarationContext::class,
+            """
+                typedef struct {
+                    logic x;
+                } s;
+            """.trimIndent(),
+            "Property(x, Nothing, SimpleDescriptor(Boolean))"
+        ) { it.findDeclaration("x") }
+    }
 }